cql plugin modification (#484)
