+++ pygeoapi/provider/geojson.py
@@ -33,9 +33,7 @@ import os
 import uuid
 
 from pygeoapi.provider.base import BaseProvider, ProviderItemNotFoundError
-from pygeoapi.util import get_filter_fields
 from pygeoapi.plugin import load_plugin
-
 
 LOGGER = logging.getLogger(__name__)
 
@@ -130,17 +128,14 @@ class GeoJSONProvider(BaseProvider):
         data = self._load()
 
         if cql_expression:
-            parser = load_plugin('cql',
-                                 {'name': 'Parser',
-                                  'cql_expression': cql_expression})
-            cql_ast = parser.create_ast()
             feature_set = data['features']
-            fields_name = list(get_filter_fields(feature_set))
-            evaluator = load_plugin('cql',
-                                    {'name': 'Evaluator',
-                                     'field_mapping': fields_name,
-                                     'mapping_choices': feature_set})
-            feature_set = evaluator.to_filter(cql_ast)
+
+            cql_handler = load_plugin('extension',
+                                      {'name': 'CQL',
+                                       'cql_expression': cql_expression,
+                                       'feature_set': feature_set})
+            feature_set = cql_handler.cql_filter()
+
             data['features'] = feature_set
 
         data['numberMatched'] = len(data['features'])
@@ -148,7 +143,7 @@ class GeoJSONProvider(BaseProvider):
         if resulttype == 'hits':
             data['features'] = []
         else:
-            data['features'] = data['features'][startindex:startindex+limit]
+            data['features'] = data['features'][startindex:startindex + limit]
             data['numberReturned'] = len(data['features'])
 
         return data
